Bug Fixed

+++ src/Pages/Welcome.js
@@ -22,26 +22,31 @@ const Welcome = () => {
     useEffect(() => {
         // let data = middlewareService.middleware({ code: code });
         // if (data.is_authenticated) localStorage.setItem("data", JSON.stringify(data.data));
-        let newObject = {
-            code: "0000",
-            address: address,
-            name: name,
-            with: withs == "1" ? "With Family" : "Self"
+        let isDataExist = localStorage.getItem("data");
+        if (!isDataExist) {
+            let newObject = {
+                code: "0000",
+                address: address ? address : "",
+                name: name ? name : "",
+                with: "",
+            };
+            if (withs && withs == "0") newObject["with"] = "Self";
+            if (withs && withs == "1") newObject["with"] = "With Family";
+            localStorage.setItem("data", JSON.stringify(newObject));
+            if (from) localStorage.setItem("from", from == "0" ? "rina" : "sourav");
         };
-        localStorage.setItem("data", JSON.stringify(newObject));
-        if (from) localStorage.setItem("from", from == "0" ? "rina" : "sourav");
         if (!isPlaying) setIsPopupOpen(true);
     }, []);
 
     return (
         <Layout>
-            <Link to="/address" className="fixed bottom-[350px] left-0" onClick={() => navigator.vibrate(75)}>
+            <Link to="/address" className="fixed bottom-[300px] left-0" onClick={() => navigator.vibrate(75)}>
                 <div className="font-playwrite p-2 mb-1 text-white shadow-md shadow-slate-500 rounded-r-full bg-black cursor-pointer select-none flex justify-center items-center gap-1 animate-blink">
                     <IoLocationOutline />
                     <p className="text-xs">Venue</p>
                 </div>
             </Link>
-            <Link to="/home" className="fixed bottom-[350px] right-0" onClick={() => navigator.vibrate(75)}>
+            <Link to="/home" className="fixed bottom-[300px] right-0" onClick={() => navigator.vibrate(75)}>
                 <div className="font-playwrite p-2 mb-1 text-white shadow-md shadow-slate-500 rounded-l-full bg-black cursor-pointer select-none flex justify-center items-center gap-1 animate-blink">
                     <p className="text-xs">Greetings</p>
                     <IoArrowForwardOutline />
@@ -51,7 +56,6 @@ const Welcome = () => {
             <div className='min-h-screen justify-center items-center flex flex-col rounded-br-3xl rounded-tl-3xl'>
                 <img alt='Swastick' className='w-[10%] text-center' src={`${process.env.PUBLIC_URL}/swastick.png`} />
                 <Divider />
-                {/* <p className='text-yellow-500 font-dancing text-2xl font-medium'> Om Shri Ganeshaya Namaha</p> */}
                 <p className='text-yellow-500 font-dancing text-4xl font-medium animate-text-from-left'>~ Welcome ~</p>
                 <Divider />
                 <PlayButton isOpen={isPopupOpen} onClose={() => setIsPopupOpen(false)} />
@@ -59,7 +63,7 @@ const Welcome = () => {
                     <img alt='Ganesh' src={`${process.env.PUBLIC_URL}/ganesh.png`} className='w-[50%] text-center' />
                 </div>
                 <Divider />
-                <p className='text-yellow-500 font-dancing text-2xl font-medium animate-text-from-right'> Om Shri Ganeshaya Namaha</p>
+                <p className='text-yellow-500 font-dancing text-2xl font-medium animate-text-from-right'>Om Shri Ganeshaya Namaha</p>
                 <Divider />
             </div>
         </Layout>
